refactor: improve commitment type decoding

+++ crates/actors/src/mempool_service/pledge_provider.rs
@@ -25,7 +25,7 @@ impl MempoolPledgeProvider {
 
 #[async_trait::async_trait]
 impl PledgeDataProvider for MempoolPledgeProvider {
-    async fn pledge_count(&self, user_address: Address) -> usize {
+    async fn pledge_count(&self, user_address: Address) -> u64 {
         // Get the canonical pledge count from the blockchain state
         let canonical_count = {
             let commitment_snapshot = self
@@ -69,8 +69,8 @@ impl PledgeDataProvider for MempoolPledgeProvider {
 
         // Calculate effective pledge count:
         // canonical pledges + pending pledges - pending unpledges
-        canonical_count
-            .saturating_add(pending_pledges)
-            .saturating_sub(pending_unpledges)
+        (canonical_count as u64)
+            .saturating_add(pending_pledges as u64)
+            .saturating_sub(pending_unpledges as u64)
     }
 }
